@@ -1,16 +1,11 @@
 // The Vue build version to load with the `import` command
 // (runtime-only or standalone) has been set in webpack.base.conf with an alias.
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App'
+
+/**添加-vue路由器 */
 import router from './router'
-import utils from './utils'
 
-Vue.config.productionTip = false
-
-/* eslint-disable no-new */
-new Vue({
-    el: '#app',
-    router,
-    components: { App },
-    template: '<App/>'
-})
+createApp(App)
+    .use(router)
+    .mount('#app')
